created a calculatorÐ

+++ app/src/main/java/com/example/coinalculator/ui/calculator/data/CalculatorEntity.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CalculatorEntity (
     val name: String,
-    val price: Long
+    val price: Float
 )
