+++ app/src/main/java/com/example/coinalculator/ui/calculator/domain/CoinCalculator.kt
@@ -2,5 +2,5 @@ package com.example.coinalculator.ui.calculator.domain
 
 data class CoinCalculator (
     val name : String,
-    val price: Long,
+    val price: Float,
 )
